requirements.txt
SydneyClient
piexif
ExifTag
pathlib